# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Trita API
  version: "2026-08-23"
  description: |
    Trita platform belső API-leírója — kézzel karbantartott, a route-fájlok
    Zod-sémáiból származtatva. Auth: Clerk session cookie (böngésző-hívások),
    kivéve a jelölt publikus, token-alapú végpontokat.

    Csoportok (tags): Public · Assessment · Observer · Profile · Notifications
    · Feedback · Team · TeamRoles/Trust/PsychSafety · Org · Manager/Hiring
    · Admin · Inquiry · System.

    Generálás: a projekt sima Next.js route-handlereket használ (nincs
    dekorátor-réteg), ezért teljesen automatikus generátor nincs bekötve.
    Ajánlott irány: zod-openapi (a meglévő Zod-sémák annotálása) vagy
    next-rest-framework — lásd docs/security/security-audit-2026-07-22.md.
servers:
  - url: https://trita.io
  - url: http://localhost:3000
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk session cookie — a böngésző automatikusan küldi.
    cronBearer:
      type: http
      scheme: bearer
      description: CRON_SECRET Bearer tokenként (Vercel cron).
    svixSignature:
      type: apiKey
      in: header
      name: svix-signature
      description: Svix webhook aláírás (svix-id + svix-timestamp + svix-signature).
...
security:
  - clerkSession: []
...
paths:
...
  /api/org/{id}/campaigns/{campaignId}:
    get: { tags: [Org], summary: Kampány részletei, responses: { "200": { description: OK } } }
    patch:
      tags: [Org]
      summary: Kampánystátusz-váltás vagy custom draft szerkesztése
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [status]
                  properties: { status: { type: string, enum: [DRAFT, ACTIVE, CLOSED] } }
                - type: object
                  required: [action]
                  properties:
                    action: { type: string, const: edit_draft }
                    types:
                      type: array
                      minItems: 1
                      maxItems: 7
                      description: Csak preset nélküli custom kampánynál módosítható.
                      items: { type: string, enum: [SELF_ASSESSMENT, OBSERVER_360, TEAM_ROLE, TEAM_ROLE_360, TRUST_360, PSYCH_SAFETY, PEER_FEEDBACK] }
                    teamId: { type: [string, "null"] }
                    teamIds: { type: array, maxItems: 50, items: { type: string } }
                    stepIntervalHours: { type: integer, minimum: 0, maximum: 168 }
                    peerFeedbackAnonymous: { type: boolean }
                    name: { type: string, maxLength: 100 }
                    description: { type: [string, "null"], maxLength: 500 }
      responses:
        "200": { description: OK }
        "409": { description: Érvénytelen állapotváltás vagy rögzített preset-lépéssor }
    post:
      tags: [Org]
      summary: Résztvevők hozzáadása
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [userIds]
              properties:
                userIds: { type: array, minItems: 1, maxItems: 50, items: { type: string } }
      responses: { "200": { description: OK } }
...